--- client.go
@@ -1,30 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"net/http"
-	"io/ioutil"
-)
-
-func main(){
-	HttpGet()
-}
-
-func HttpGet(){
-	tr := &http.Transport{}
-	client := http.Client{
-		Transport: tr,
-	}
-	req, _ := http.NewRequest("GET", "http://127.0.0.1:8000/ping", nil)
-	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-	resp, err := client.Do(req)
-	if err != nil{
-		return
-	}
-	defer resp.Body.Close()
-	bodyBytes, err := ioutil.ReadAll(resp.Body)
-	if err != nil{
-		return
-	}
-	fmt.Println(string(bodyBytes))
-}
