+++ client/build.sh
@@ -1,0 +1,4 @@
+RUN_NAME="client"
+mkdir -p output/bin
+export GO111MODULE=on
+go build  -o output/bin/${RUN_NAME}
