+++ client/go.mod
@@ -1,0 +1,3 @@
+module github.com/Http/client
+
+go 1.15
