+++ server/build.sh
@@ -1,0 +1,6 @@
+RUN_NAME="server"
+
+mkdir -p output/bin
+export GO111MODULE=on
+
+go build  -o output/bin/${RUN_NAME}
