+++ server/go.mod
@@ -1,0 +1,3 @@
+module github.com/Http/server
+
+go 1.15
